Training Projects/Ollama Recruiting Agency/requirements.txt
swarms == 0.1.0
openai == 1.12.0
pdfminer.six==20221105
python-dotenv==1.0.0
rich==13.7.0
streamlit==1.32.0
streamlit-extras==0.4.0
streamlit-option-menu==0.3.12